DON'T UPDATE UNLESS YOU WANT TO TEST THIS
I’m testing this code on my computer, however it keeps crashing. I am
unsure and can not figure it out… It’s after I finish writing the
TestIns class. I’m going to be looking online probably, but wanted to
see if the project worked on y’alls computer.